+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/ImageFormat.cs
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: 2025 Unity Technologies and the glTFast authors
+// SPDX-License-Identifier: Apache-2.0
+
+using System;
+
+namespace GLTFast
+{
+    enum ImageFormat
+    {
+        Unknown,
+        PNG,
+        Jpeg,
+        Ktx
+    }
+}